Made chrono optional
