+++ .github/workflows/circadian-tools-build.yml
@@ -36,4 +36,4 @@ jobs:
     - name: Run cargo test
       uses: actions-rs/cargo@v1
       with:
-        command: test
+        command: test --all-features
